Interesting feature (might be useful for you Ivan):
factors now can have an order number assigned to within a solver. This order number defines when to start checking for and adding formulas to the solving process. For example, it makes sense to give hard LEQ constraints a low number

set solver.order.atMostOneSlot = 0

and soft constraints a higher order

set solver.order.slotPair = 1;

When the solver runs, it will first try to fix all atMostOneSlot violations before it goes on to the slotPair factors. 

By default the order of a factor is 0. Note that you can use any number (such as 1000) for the order of a factor. You don't need to have all orders from 0 to 999 defined. 

This is similar to grounding in advance, but allows a bit more control.