new type feature: Now types do not need to be fully defined in advance. Just writing
type Word: ...;

will be enough. When new data loaded (from a corpus) types created with ... in front will be filled up on the fly. However, if you train a model using these types and plan to reuse it later you need to save the types by simply writing

types to "filename.pml";

This creates a pml file with type definitions. At test time just include this file as first command:

include "filename.pml"; 

And that should work. However, the old way of doings things in advance still works, too.

This has another advantage: even at test time the ... types are further extended and it will be possible to see the actual data instead of UNKNOWN tokens.