- srl features - unquantified formulas: sometimes one wants to define factors that use no quantification variables. For example, if I want to define a factor that forbids more than 2 TOLOCs per sentence I would like to write
factor: |Int t: slot(t,TOLOC)| <= 2;

but since a factor is defined as a sum over an index set, for this to work one would have had to define a dummy type with a single constant (type Singleton: "somerandomconstant";} and then

factor: for Singleton s:  |Int t: slot(t,TOLOC)| <= 2;

(that is, for each element in Singleton, apply the formula, and since there is only one element in Singleton it does what we want). 

However, now one can use the first, more compact version.